# 🚀 Speed up toMetricTagsID