`toMetricTagsID` is called every time a metric is incremented/changed. This PR updates the implementation to run faster and allocate less memory.

Benchmark results:
Old
```
goos: darwin
goarch: amd64
pkg: github.com/palantir/pkg/metrics
BenchmarkRegisterMetric/1_tag-8          2000000     908 ns/op     800 B/op   10 allocs/op
BenchmarkRegisterMetric/10_tag-8          300000    4905 ns/op    2910 B/op   27 allocs/op
BenchmarkRegisterMetric/100_tag-8          30000   53716 ns/op   30639 B/op  136 allocs/op
PASS
ok  	github.com/palantir/pkg/metrics	6.802s
```
New
```
goos: darwin
goarch: amd64
pkg: github.com/palantir/pkg/metrics
BenchmarkRegisterMetric/1_tag-8          5000000     300 ns/op     208 B/op    4 allocs/op
BenchmarkRegisterMetric/10_tag-8         2000000     657 ns/op     368 B/op    4 allocs/op
BenchmarkRegisterMetric/100_tag-8         200000    8304 ns/op    2704 B/op    4 allocs/op
PASS
ok  	github.com/palantir/pkg/metrics	5.880s
```

This was flagged by an internal team looking at the pprofs of their service.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/pkg/130)
<!-- Reviewable:end -->